fix: clean up unused params and make TotalCount long to avoid overflow
Agent-Logs-Url: https://github.com/ivanopaulon/EventForge/sessions/b82031d8-a15c-44fb-bed0-9a148896ae11

+++ EventForge.Server/Pages/Dashboard/Logs.cshtml.cs
@@ -15,7 +15,7 @@ namespace EventForge.Server.Pages.Dashboard;
 public class LogsModel(ILogManagementService logManagementService, ILogger<LogsModel> logger) : PageModel
 {
     public List<SystemLogDto> Logs { get; set; } = [];
-    public int TotalCount { get; set; }
+    public long TotalCount { get; set; }
     public int TotalPages { get; set; }
     public int PageSize { get; set; } = 50;
 
@@ -50,7 +50,7 @@ public class LogsModel(ILogManagementService logManagementService, ILogger<LogsM
             var result = await logManagementService.GetApplicationLogsAsync(queryParams, HttpContext.RequestAborted);
 
             Logs = result.Items?.ToList() ?? [];
-            TotalCount = (int)result.TotalCount;
+            TotalCount = result.TotalCount;
             TotalPages = (int)Math.Ceiling(TotalCount / (double)PageSize);
             CurrentPage = Math.Max(1, Math.Min(CurrentPage, Math.Max(1, TotalPages)));
         }
